+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [2.6.1 - 2019-08-10]
+
+### Fixed
+
 - Added missing indices on `crawls` and `crawl_interactions` tables.
 - Added table to store most recent crawl. This speeds up the instance view by a lot!